+++ src/editing.ts
@@ -4,13 +4,15 @@ import { NEURO } from '@/constants';
 import { DiffRangeType, escapeRegExp, getDiffRanges, getFence, getPositionContext, getProperty, getVirtualCursor, showDiffRanges, isPathNeuroSafe, logOutput, NeuroPositionContext, setVirtualCursor, simpleFileName, substituteMatch, clearDecorations } from '@/utils';
 import { ActionData, actionValidationAccept, actionValidationFailure, ActionValidationResult, RCEAction, contextFailure, stripToActions, actionValidationRetry, contextNoAccess } from '@/neuro_client_helper';
 import { PERMISSIONS, getPermissionLevel, CONFIG, isActionEnabled } from '@/config';
+import { createCursorPositionChangedEvent } from '@events/cursor';
+import { RCECancelEvent } from '@events/utils';
 import type { JSONSchema7 } from 'json-schema';
 
 const CONTEXT_NO_ACCESS = 'You do not have permission to access this file.';
 const CONTEXT_NO_ACTIVE_DOCUMENT = 'No active document to edit.';
 
 type MatchOptions = 'firstInFile' | 'lastInFile' | 'firstAfterCursor' | 'lastBeforeCursor' | 'allInFile';
-const MATCH_OPTIONS: string[] = ['firstInFile', 'lastInFile', 'firstAfterCursor', 'lastBeforeCursor', 'allInFile'] as const;
+const MATCH_OPTIONS: MatchOptions[] = ['firstInFile', 'lastInFile', 'firstAfterCursor', 'lastBeforeCursor', 'allInFile'] as const;
 const POSITION_SCHEMA: JSONSchema7 = {
     type: 'object',
     properties: {
@@ -166,6 +168,29 @@ function createLineRangeValidator(path = '') {
     };
 }
 
+const cancelOnDidChangeTextDocument = () => new RCECancelEvent({
+    reason: 'the active document was changed.',
+    events: [
+        [vscode.workspace.onDidChangeTextDocument, null],
+    ],
+});
+const cancelOnDidChangeActiveTextEditor = () => new RCECancelEvent({
+    reason: 'you\'ve switched files.',
+    events: [
+        [vscode.window.onDidChangeActiveTextEditor, null],
+    ],
+});
+
+const commonCancelEvents: (() => RCECancelEvent)[] = [
+    cancelOnDidChangeTextDocument,
+    cancelOnDidChangeActiveTextEditor,
+];
+
+const commonCancelEventsWithCursor: (() => RCECancelEvent)[] = [
+    ...commonCancelEvents,
+    createCursorPositionChangedEvent,
+];
+
 export const editingActions = {
     place_cursor: {
         name: 'place_cursor',
@@ -173,7 +198,8 @@ export const editingActions = {
         schema: POSITION_SCHEMA,
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handlePlaceCursor,
-        validator: [checkCurrentFile, createPositionValidator()],
+        validators: [checkCurrentFile, createPositionValidator()],
+        cancelEvents: commonCancelEvents,
         promptGenerator: (actionData: ActionData) => `${actionData.params.type === 'absolute' ? 'place the cursor at' : 'move the cursor by'} (${actionData.params.line}:${actionData.params.column}).`,
     },
     get_cursor: {
@@ -181,7 +207,8 @@ export const editingActions = {
         description: 'Get the current cursor position and the text surrounding it.',
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleGetCursor,
-        validator: [checkCurrentFile],
+        validators: [checkCurrentFile],
+        cancelEvents: commonCancelEventsWithCursor,
         promptGenerator: 'get the current cursor position and the text surrounding it.',
     },
     get_content: {
@@ -189,8 +216,11 @@ export const editingActions = {
         description: 'Get the contents of the current file.',
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleGetContent,
+        cancelEvents: [
+            cancelOnDidChangeTextDocument,
+        ],
         promptGenerator: 'get the current file\'s contents.',
-        validator: [checkCurrentFile],
+        validators: [checkCurrentFile],
     },
     insert_text: {
         name: 'insert_text',
@@ -211,7 +241,13 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleInsertText,
-        validator: [checkCurrentFile, createPositionValidator('position'), createStringValidator(['text'])],
+        cancelEvents: [
+            ...commonCancelEvents,
+            (actionData: ActionData) => {
+                return actionData.params.position ? null : createCursorPositionChangedEvent();
+            },
+        ],
+        validators: [checkCurrentFile, createPositionValidator('position'), createStringValidator(['text'])],
         promptGenerator: (actionData: ActionData) => {
             const lineCount = actionData.params.text.trim().split('\n').length;
             let text = `insert ${lineCount} line${lineCount === 1 ? '' : 's'} of code`;
@@ -234,6 +270,7 @@ export const editingActions = {
         name: 'insert_lines',
         description: 'Insert code below a certain line.'
             + ' Defaults to your current cursor\'s location'
+            + ' The insertUnder parameter is one-based, not zero-based.'
             + ' Remember to add indents after newlines where appropriate.'
             + ' Your cursor will be moved to the end of the inserted line.', // TODO: Clarify cursor stuff again
         schema: {
@@ -252,7 +289,13 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleInsertLines,
-        validator: [checkCurrentFile, createStringValidator(['lines'])],
+        cancelEvents: [
+            ...commonCancelEvents,
+            (actionData: ActionData) => {
+                return actionData.params.position ? null : createCursorPositionChangedEvent();
+            },
+        ],
+        validators: [checkCurrentFile, createStringValidator(['lines'])],
         promptGenerator: (actionData: ActionData) => {
             const lines = actionData.params.text.trim().split('\n').length;
             const insertUnder = actionData.params.insertUnder;
@@ -278,7 +321,8 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleReplaceText,
-        validator: [checkCurrentFile, createStringValidator(['find', 'replaceWith']), createLineRangeValidator('lineRange')],
+        cancelEvents: [cancelOnDidChangeActiveTextEditor],
+        validators: [checkCurrentFile, createStringValidator(['find', 'replaceWith']), createLineRangeValidator('lineRange')],
         promptGenerator: (actionData: ActionData) => {
             let text = 'replace ';
             const target = actionData.params.useRegex ? escapeRegExp(actionData.params.find) : actionData.params.find;
@@ -332,7 +376,8 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleDeleteText,
-        validator: [checkCurrentFile, createStringValidator(['find']), createLineRangeValidator('lineRange')],
+        cancelEvents: [cancelOnDidChangeActiveTextEditor],
+        validators: [checkCurrentFile, createStringValidator(['find']), createLineRangeValidator('lineRange')],
         promptGenerator: (actionData: ActionData) => {
             let text = 'delete ';
             const target = actionData.params.useRegex ? escapeRegExp(actionData.params.find) : actionData.params.find;
@@ -392,7 +437,8 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleFindText,
-        validator: [checkCurrentFile, createStringValidator(['find']), createLineRangeValidator('lineRange')],
+        cancelEvents: [cancelOnDidChangeActiveTextEditor],
+        validators: [checkCurrentFile, createStringValidator(['find']), createLineRangeValidator('lineRange')],
         promptGenerator: (actionData: ActionData) => {
             let text = 'find ';
             const target = actionData.params.useRegex ? escapeRegExp(actionData.params.find) : actionData.params.find;
@@ -440,7 +486,8 @@ export const editingActions = {
             + ' If this doesn\'t work, tell Vedal to focus your VS Code window.',
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleUndo,
-        validator: [checkCurrentFile],
+        cancelEvents: commonCancelEvents,
+        validators: [checkCurrentFile],
         promptGenerator: 'undo the last action.',
     },
     save: {
@@ -448,7 +495,16 @@ export const editingActions = {
         description: 'Manually save the currently open document.',
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleSave,
-        validator: [checkCurrentFile],
+        cancelEvents: [
+            ...commonCancelEvents,
+            () => new RCECancelEvent({
+                reason: 'the active document was saved.',
+                events: [
+                    [vscode.workspace.onDidSaveTextDocument, null],
+                ],
+            }),
+        ],
+        validators: [checkCurrentFile],
         promptGenerator: 'save.',
     },
     rewrite_all: {
@@ -464,7 +520,8 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleRewriteAll,
-        validator: [checkCurrentFile, createStringValidator(['content'])],
+        cancelEvents: [cancelOnDidChangeActiveTextEditor],
+        validators: [checkCurrentFile, createStringValidator(['content'])],
         promptGenerator: (actionData: ActionData) => {
             const lineCount = actionData.params.content.trim().split('\n').length;
             return `rewrite the entire file with ${lineCount} line${lineCount === 1 ? '' : 's'} of content.`;
@@ -486,7 +543,8 @@ export const editingActions = {
         },
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleRewriteLines,
-        validator: [checkCurrentFile, createLineRangeValidator(), createStringValidator(['content'])],
+        cancelEvents: commonCancelEvents,
+        validators: [checkCurrentFile, createLineRangeValidator(), createStringValidator(['content'])],
         promptGenerator: (actionData: ActionData) => {
             const lineCount = actionData.params.content.trim().split('\n').length;
             return `rewrite lines ${actionData.params.startLine}-${actionData.params.endLine} with ${lineCount} line${lineCount === 1 ? '' : 's'} of content.`;
@@ -500,7 +558,8 @@ export const editingActions = {
         schema: LINE_RANGE_SCHEMA,
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleDeleteLines,
-        validator: [checkCurrentFile, createLineRangeValidator()],
+        cancelEvents: commonCancelEvents,
+        validators: [checkCurrentFile, createLineRangeValidator()],
         promptGenerator: (actionData: ActionData) => {
             return `delete lines ${actionData.params.startLine}-${actionData.params.endLine}.`;
         },
@@ -514,7 +573,8 @@ export const editingActions = {
         schema: LINE_RANGE_SCHEMA,
         permissions: [PERMISSIONS.editActiveDocument],
         handler: handleHighlightLines,
-        validator: [checkCurrentFile, createLineRangeValidator()],
+        cancelEvents: commonCancelEvents,
+        validators: [checkCurrentFile, createLineRangeValidator()],
         promptGenerator: (actionData: ActionData) => `highlight lines ${actionData.params.startLine}-${actionData.params.endLine}.`,
     },
 } satisfies Record<string, RCEAction>;