+++ src/neuro_client_helper.ts
@@ -6,6 +6,7 @@ import { Action } from 'neuro-game-sdk';
 import { Permission, PermissionLevel } from '@/config';
 import { logOutput } from '@/utils';
 import { PromptGenerator } from '@/rce';
+import { RCECancelEvent } from '@events/utils';
 import { JSONSchema7 } from 'json-schema';
 
 /** Data used by an action handler. */
@@ -39,7 +40,15 @@ export interface RCEAction extends TypedAction {
     /** The permissions required to execute this action. */
     permissions: Permission[];
     /** The function to validate the action data *after* checking the schema. */
-    validator?: (((actionData: ActionData) => ActionValidationResult) | ((actionData: ActionData) => Promise<ActionValidationResult>))[];
+    validators?: (((actionData: ActionData) => ActionValidationResult) | ((actionData: ActionData) => Promise<ActionValidationResult>))[];
+    /**
+     * Cancellation events attached to the action that will be automatically set up.
+     * Each cancellation event will be setup in parallel to each other.
+     * If one cancellation event fires, the request is cancelled and all listeners will be disposed as soon as possible.
+     * 
+     * Following VS Code's pattern, Disposables will not be awaited if async.
+     */
+    cancelEvents?: ((actionData: ActionData) => RCECancelEvent | null)[];
     /** The function to handle the action. */
     handler: (actionData: ActionData) => string | undefined;
     /**