+++ CHANGELOG.md
@@ -16,10 +16,12 @@ Changes between each version before then will not be listed.
   - `autoConnect`, which controls whether or not NeuroPilot will auto-connect to the Neuro API on startup.
   - `retryAmount`, which dictates how many times NeuroPilot will attempt to *re*connect to the Neuro API. The _total_ amount of tries will be this value + 1. 
   - `retryInterval`, which controls the amount of time to wait between each retry.
+- `neuropilot.enableCancelEvents` - Whether or not to enable cancel events. See below for an explainer.
 
 ### New commands
 
 - `neuropilot.disconnect` - Disconnects from the Neuro API without reconnecting.
+- `neuropilot.sendSelectionToNeuro` - Sends the current cursor selection to Neuro.
 
 ### Added features
 
@@ -27,6 +29,10 @@ Changes between each version before then will not be listed.
 - A tooltip will now show when hovering over Neuro-highlighted text, indicating that Neuro highlighted it, and how (either through finding text or manually).
 - There is now an option to send the current cursor selection to Neuro. This can be invoked either by command or through right-click context menu. These options will only appear when *your cursor* is highlighting code.
 - There is a new detector ran on startup to look for deprecated settings and if found, will ask you to migrate them. You can choose to not show that notice ever again by selecting the "Don't show again" option. This persists across sessions.
+- The RCE framework now has a new component: **cancellation events**.
+  - This automatically cancels Neuro's action requests if certain events happen.
+  - Example: If Neuro wants to insert text, that request will auto-cancel if the active file is switched.
+  - You can disable this with the new `Enable Cancel Events` setting (defaulted to on).
 
 ### Changes